Add test for placement method
