+++ lib/board.rb
@@ -7,13 +7,13 @@ class Board
     puts " #{@board_arr[6]} | #{@board_arr[7]} | #{@board_arr[8]}"
   end
 
-  def placement(user_move)
-    if user_move == @user_one_move
+  def placement(user_move, player)
+    if player == 1
       user_symbol = 'X'
-      @board_arr[@user_one_move - 1] = user_symbol
-    elsif user_move == @user_two_move
+      @board_arr[user_move - 1] = user_symbol
+    elsif player == 2
       user_symbol = 'O'
-      @board_arr[@user_two_move - 1] = user_symbol
+      @board_arr[user_move - 1] = user_symbol
     end
   end
 end
